+++ front-end/src/App.js
@@ -5,6 +5,7 @@ import Home from './Home'
 import Results from './Results'
 import ClassInfo from './ClassInfo';
 import Account from './Account';
+import Login from './Login';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 // import Dropdown from './Dropdown'
 
@@ -36,6 +37,7 @@ function App() {
           <Route  exact path="/results" component={Results} />
           <Route  exact path="/ClassInfo" component={ClassInfo} />
           <Route  exact path="/Account" component={Account} />
+          <Route exact path="/Login" component={Login} />
 
           {/* <Route path="/">
             <Home />
