+++ front-end/src/Login.css
@@ -1,38 +1,26 @@
 .Login{
-    position: relative;
-}
-.inputs{
-    align-self: center;
-    border: 3px solid #7851A9;
-    width: 38rem;
-    height: 3rem;
-    color: gray;
-    background-color: white;
     text-align: center;
-    align-content: center;
-    font-size: 20px;
-
-}
-
-.submit-button{
-    background-color: teal;
-    border: solid;
-    border-radius: 10px;
-    color: white;
-    padding: 12px 16px;
-    font-size: 30px;
-    cursor: pointer;
-    position: relative;
-    margin-left: 5rem;
-
-}
-.submit-button:hover {
-    background-color: #360358;
-}
-
-.center{
-    justify-self: center;
-    align-self: center;
+    white-space: pre-wrap;
     align-content: center;
     align-items: center;
+    vertical-align: center;
+}
+
+.Login-page{
+    position: relative;
+    border-style: solid;
+    border-color: white;
+    border:10px;
+    background-color: #57068C;
+    height: 70rem;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
+    white-space: pre-wrap;
+
+    
 }
