+++ front-end/src/Login.js
@@ -1,98 +1,31 @@
-import React,{useState} from'react';
-import ReactDOM from 'react-dom';
+import logo from './logo.svg';
+import React from 'react';
 import MenuBar from './MenuBar';
-import './Login.css'
 import './MenuBar.css';
-import { Link } from 'react-router-dom';
-import Account from './Account';
-import './Account.css';
+import ReactDOM from 'react-dom';
+import App from './App'
+import './Login.css'
+import LoginLogout from'./LoginLogout'
+import { Link, Switch } from 'react-router-dom';
 
-const Login=(props)=> {
+const Login=(props)=>{
+      return(
+        <div className="Login-page">
+            <MenuBar>
 
-    const [email,setEmail]=useState("Enter your Email");
-    const [password,setPassword]=useState("Enter your password");
+            </MenuBar>  
+            <p>
+              Waitlisted?
+            </p>
+            <LoginLogout/>
 
-/* function submitLogin(e) {
-    axios 
-        .post("/login", loginDeets)
-        .then((response) => {
-            console.log("login response ", response.data);
-
-            if (!response.data.error) {
-                console.log("login success");
-            } else {
-                setError(true);
-            }
-        })
-        .catch((err) => {
-            console.log("Error", err);
-        });
-}*/
-
-    const handleChange = ({ target }) => { 
-      setEmail( target.value );
-    };
-    const handlePassword=({target})=>{
-        setPassword(target.value);
-    };
-    const handleClickEmail = (e) => {
-      setEmail('');
-    };
-    const handleClickPassword = () => {
-      setPassword('');
-    };
-    
-      return (
-        <React.Fragment>
-          {}
-          <MenuBar>
-          </MenuBar>
-            <form className="box">
-              <p>
-                {}
-                  <input className="inputs"
-                  type="email"
-                  name="email"
-                  value={email}
-
-                  onChange={handleChange}
-                  onClick={handleClickEmail}
-                  />
-              </p>
             
-              <p> 
-              {}
-                  <input className="inputs"
-                  type="password"
-                  name="password"
-                  value={password}
+          {/* </header> */}
+          
+        </div>
+      )
+      
+      
+};
 
-                  onChange={handlePassword}
-                  onClick={handleClickPassword}
-                  />
-              </p>
-            
-              {}
-
-              <Link to="/Account">
-                <center>
-                 <button className="submit-button"> Login </button>
-                </center>
-                
-              </Link>
-
-              <Link to="/Account">
-                <center>
-                 <button className="submit-button"> Signup </button>
-                </center>
-                
-              </Link>
-
-            </form>   
-          {}
-        </React.Fragment>
-
-      );    
-   }
-
-export default Login
+export default Login;
