+++ front-end/src/LoginLogout.css
@@ -1,0 +1,38 @@
+.LoginLogout{
+    position: relative;
+}
+.inputs{
+    align-self: center;
+    border: 3px solid #7851A9;
+    width: 38rem;
+    height: 3rem;
+    color: gray;
+    background-color: white;
+    text-align: center;
+    align-content: center;
+    font-size: 20px;
+
+}
+
+.submit-button{
+    background-color: teal;
+    border: solid;
+    border-radius: 10px;
+    color: white;
+    padding: 12px 16px;
+    font-size: 30px;
+    cursor: pointer;
+    position: relative;
+    margin-left: 5rem;
+
+}
+.submit-button:hover {
+    background-color: #360358;
+}
+
+.center{
+    justify-self: center;
+    align-self: center;
+    align-content: center;
+    align-items: center;
+}
